backend_app/requirements.txt

django
djangorestframework
django-filter
django-cors-headers
djangorestframework-simplejwt
Pillow

langchain
langchain_community
langgraph

#local
langchain-ollama